change routes, rename folder +...
