+++ client/src/app/+devices-setup/blinds-setup/blinds-setup.component.ts
