+++ client/src/app/+devices-setup/blinds-setup/index.ts
