+++ client/src/app/+devices-setup/devices-setup.component.html
