+++ client/src/app/+devices-setup/devices-setup.module.ts
