+++ client/src/app/+devices-setup/humidity-setup/humidity-setup.component.scss
