+++ client/src/app/+devices-setup/index.ts
