+++ client/src/app/+devices-setup/temperature-setup/index.ts
