+++ client/src/app/+devices-setup/temperature-setup/temperature-setup.component.html
