+++ client/src/app/+devices-setup/temperature-setup/temperature-setup.component.ts
