+++ client/src/app/app.module.ts
@@ -14,7 +14,7 @@ import {UsersComponent} from "./users/users.component";
 import {AuthGuard} from "./auth/auth-guard.service";
 import {AuthHttp, AuthConfig} from "angular2-jwt";
 import {TemperatureViewComponent} from "./temperature-view/temperature-view.component";
-import {DevicesSetupComponent} from "./devices-setup/devices-setup.component";
+import {DevicesSetupComponent} from "./+devices-setup/devices-setup.component";
 
 
 // TODO: in welches File müsste diese Methode?
