+++ client/src/app/dashboard/dashboard.component.ts
@@ -75,5 +75,4 @@ export class DashboardComponent implements OnInit {
         break;
     }
   }
-
 }
